+++ ocpi221/common.go
@@ -144,8 +144,8 @@ const (
 )
 
 type Price struct {
-	ExclVat float64 `json:"excl_vat"`
-	InclVat float64 `json:"incl_vat"`
+	ExclVat float32 `json:"excl_vat"`
+	InclVat float32 `json:"incl_vat"`
 }
 
 type CDRToken struct {
@@ -174,7 +174,7 @@ type GeoLocation struct {
 
 type Dimension struct {
 	Type   DimensionType `json:"type"`
-	Volume float64       `json:"volume"`
+	Volume float32       `json:"volume"`
 }
 
 type ChargingPeriod struct {
@@ -185,12 +185,12 @@ type ChargingPeriod struct {
 
 type EnvironmentalImpact struct {
 	Category EnvironmentalImpactCategory `json:"category"`
-	Amount   float64                     `json:"amount"`
+	Amount   float32                     `json:"amount"`
 }
 
 type EnergySource struct {
 	Source     EnergySourceCategory `json:"source"`
-	Percentage float64              `json:"percentage"`
+	Percentage float32              `json:"percentage"`
 }
 
 type EnergyMix struct {
